fix: handle flush error when leader tries to commit
All flush() commands in the journal module now can throw a checked
exception FlushException, in order to make sure that all possible
exceptions thrown when flushing are handled.
A new class CheckedJournalException has been created, since
JournalException extends RuntimeException.

When the leader fails to flush, it resets its commitIndex to the
previous value and transitions to the Follower role.

closes #15463